Refactor CLI storage helpers into modules
